server-old/requirements.txt
dbutils==3.1.2
IPy==1.01
PyYAML==6.0.3
requests==2.32.5
tenacity==9.1.2
sqlparse==0.5.3
pyinstaller==6.16.0
GitPython==3.1.45

git+ssh://devops.aishu.cn:22/AISHUDevOps/One-Architecture/_git/proton-rds-sdk-py@release/2.0.0
